trial1-master/trial1-master/requirements.txt
pandas
dnspython==1.16.0
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
cffi==1.14.0
chardet==3.0.4
discord.py-self==1.9.2
idna==2.9
multidict==4.7.6
pycparser==2.20
PyNaCl==1.3.0
python-dotenv==0.13.0
six==1.14.0
websockets==8.1
yarl==1.4.2
youtube-dl==2020.5.3
requests
